chore: move item structs and impl to lib.rs
fmt/clippy
