+++ src/memory/inventory_manager.rs
@@ -81,12 +81,10 @@ impl UnitySerializableDictKey for InventoryItemName {
     fn read(process: &Process, item_ptr: u64) -> Result<Self, MemoryError> {
         if let Ok(guid) = process.read::<ArrayWString<128>>(item_ptr + 0x14) {
             match String::from_utf16(guid.as_slice()) {
-                Ok(value) => {
-                    match all_items().get(value.as_str()) {
-                        Some(item) => Ok(InventoryItemName(item.name.to_string())),
-                        None => Ok(InventoryItemName("UNKNOWN ITEM NAME".to_string()))
-                    }
-                }
+                Ok(value) => match all_items().get(value.as_str()) {
+                    Some(item) => Ok(InventoryItemName(item.name.to_string())),
+                    None => Ok(InventoryItemName("UNKNOWN ITEM NAME".to_string())),
+                },
                 Err(_) => Err(MemoryError::Unset),
             }
         } else {
